+++ decidim-core/spec/controllers/newsletters_controller_spec.rb
@@ -64,7 +64,7 @@ module Decidim
           let(:sent_at_time) { Time.zone.at(decrypted_string.split("-").second.to_i) }
 
           context "and newsletter notifications is true" do
-            let!(:user) { create(:user, organization: organization, id: user_id, newsletter_notifications_at: Time.zone.now) }
+            let!(:user) { create(:user, organization: organization, id: user_id, newsletter_notifications_at: Time.current) }
 
             it "unsubscribe user" do
               get :unsubscribe, params: { u: encryptor.sent_at_encrypted(user_id, time) }